Vin README.md
Manveen Kaur ’21 student commited to learning, developing skills in data science and team contribution. At Lovely Professional University, PB, Manveen learned the importance of data in business, that is essential for businesses to grow and prosper. Recently she completed applied data science capstone project. Manveen is currently finalising Bachelor’s in Computer Application studies with a concentration in Data Science and hopes to intern in an organisation in the near future.